small fix per dan's feedback
